cryptohome: add static_casts to appease `-Wformat`
`-Wformat` will soon start complaining if a value of `enum class` type
is passed into a `printf`-like function with a `%d` format specifier.
Explicitly casting to an int silences this warning.

BUG=b:305714217
TEST=emerge-jacuzzi cryptohome

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4942795
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>

+++ cryptohome/cryptorecovery/inclusion_proof.cc
@@ -342,9 +342,10 @@ std::string GetOnboardingMetadataDiff(const OnboardingMetadata& expected,
                                       const OnboardingMetadata& actual) {
   std::vector<std::string> messages;
   if (expected.cryptohome_user_type != actual.cryptohome_user_type) {
-    messages.push_back(base::StringPrintf(
-        "cryptohome_user_type: expected '%d' but got '%d'",
-        expected.cryptohome_user_type, actual.cryptohome_user_type));
+    messages.push_back(
+        base::StringPrintf("cryptohome_user_type: expected '%d' but got '%d'",
+                           static_cast<int>(expected.cryptohome_user_type),
+                           static_cast<int>(actual.cryptohome_user_type)));
   }
   if (expected.cryptohome_user != actual.cryptohome_user) {
     messages.push_back(base::StringPrintf(
